\name{partitions}
\alias{partitions}
\alias{genPL}
\alias{genCoefs}
\alias{partitionToCoef}
\alias{secondbiggest}

\title{
Functions to generate partitions and coefficients for investigating the origins of duplicate reads.
}
\description{
Functions to generate lists of partitions and coefficients, and to convert one to the other.
}
\usage{
genPL(LPL = 20)
genCoefs(PL)
partitionToCoef(x)
secondbiggest(y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{LPL}{The maximum size of a set of fragments for which to work out all possible partitions of fragments from original molecules}
  \item{PL}{A partition list generated by genPL}
  \item{x}{A specific partition}
  \item{y}{A list of numbers}
}
\value{
\item{genPL }{generates a list, the ith item of which is a set of possible partitions of i fragments}
\item{genCoefs }{generates a list, the ith item of which is a set of coefficients associated with a set of partitions in the input}
\item{partitionToCoef }{generates a a coefficient to accompany the input partition}
\item{secondbiggest }{returns the second largest component of the input vector}
}
\references{
Lynch Smith Eldridge and Tavare 2016
}
\author{
Andy Lynch
}

\examples{
# PL<-genPL(10)
# Coefs<-genCoefs(PL)
}